refactor: modifica o nome e a estrutura do projeto
