+++ api-backend/src/main/java/br/com/ReterMais/ReterMaisApplication.java
@@ -1,13 +1,13 @@
-package br.com.TelcoCustomerChurn;
+package br.com.ReterMais;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class TelcoCustomerChurnApplication {
+public class ReterMaisApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(TelcoCustomerChurnApplication.class, args);
+		SpringApplication.run(ReterMaisApplication.class, args);
 	}
 
 }
